Refactored email exception package to singular noun, was exceptions
This change is to make the package consistent with the rest of the packages which are using singular nouns
