+++ dces-report-service/src/main/java/uk/gov/justice/laa/crime/dces/report/utils/email/EmailClient.java
@@ -1,7 +1,7 @@
 package uk.gov.justice.laa.crime.dces.report.utils.email;
 
-import uk.gov.justice.laa.crime.dces.report.utils.email.exceptions.EmailClientException;
-import uk.gov.justice.laa.crime.dces.report.utils.email.exceptions.EmailObjectInvalidException;
+import uk.gov.justice.laa.crime.dces.report.utils.email.exception.EmailClientException;
+import uk.gov.justice.laa.crime.dces.report.utils.email.exception.EmailObjectInvalidException;
 
 public interface EmailClient {
     void send(EmailObject emailObject) throws EmailClientException, EmailObjectInvalidException;
