+++ dces-report-service/src/main/java/uk/gov/justice/laa/crime/dces/report/utils/email/exception/EmailClientException.java
@@ -1,4 +1,4 @@
-package uk.gov.justice.laa.crime.dces.report.utils.email.exceptions;
+package uk.gov.justice.laa.crime.dces.report.utils.email.exception;
 
 public class EmailClientException extends RuntimeException {
     public EmailClientException(String message) {
