+++ dces-report-service/src/main/java/uk/gov/justice/laa/crime/dces/report/utils/email/exception/EmailObjectInvalidException.java
@@ -1,4 +1,4 @@
-package uk.gov.justice.laa.crime.dces.report.utils.email.exceptions;
+package uk.gov.justice.laa.crime.dces.report.utils.email.exception;
 
 public class EmailObjectInvalidException extends RuntimeException {
     public EmailObjectInvalidException(String message) {
